XC-293: Re-organize sol_rpc_types into sub-modules
